refactor: move cookie, metadata and mock-server code out of main.rs
main.rs mixed CLI parsing, cookie handling, URL helpers and the test
mock server in one file. The domain logic now lives in the lib;
main.rs only orchestrates the CLI.

- Extract the Cookie header building (raw string or Netscape
  cookies.txt, with domain/path/expiry/secure filtering) into a new
  cookie module, tests included
- Move the URL helpers (get_xml_url, encode_download_path,
  xml_file_name_of) and the metadata path (is_url_accessible,
  fetch_xml_metadata) into archive_metadata; split fetch_and_parse_xml
  out so tests can point it at the local mock server
- Extract handle_existing_file and install_downloaded_file from
  process_file, and expose parse_last_modified / sync_file_mtime so
  the saved _files.xml gets the same mtime handling
- Move the in-memory HTTP mock server from the downloader tests into
  the shared test_support module (now also used by the
  archive_metadata tests), with request-method tracking and HEAD
  support for the accessibility check
- Fix the "Downloaded" tree glyph printing uncoloured:
  print_downloaded_line took the pre-styled prefix as &str, but a
  &ColoredString coerces through Deref to the plain, un-styled text.
  It now takes the ColoredString by value
- Refresh the AGENTS.md architecture section (new cookie module,
  corrected one-liners)

All 60 unit/doc tests, clippy and fmt pass.
