+++ src/lib.rs
@@ -7,6 +7,7 @@
 
 pub mod archive_metadata;
 pub mod constants;
+pub mod cookie;
 pub mod downloader;
 pub mod error;
 #[cfg(test)]
